feat: streamorder as integer output and not float

+++ src/earthkit/hydro/streamorder/array/_operations.py
@@ -57,10 +57,10 @@ def flow_strahler(
 @multi_backend(jax_static_args=["xp", "river_network", "return_type"])
 def strahler(xp, river_network, return_type):
 
-    field = xp.zeros(river_network.n_nodes)
+    field = xp.zeros(river_network.n_nodes, dtype=int)
     # TODO: make xp-agnostic
     field[river_network.sources] = 1
-    counts = xp.zeros(river_network.n_nodes)
+    counts = xp.zeros(river_network.n_nodes, dtype=int)
 
     decorated_func = mask(return_type == "gridded")(flow_strahler)
     return decorated_func(xp, river_network, field, counts)
@@ -68,7 +68,7 @@ def strahler(xp, river_network, return_type):
 
 @multi_backend(jax_static_args=["xp", "river_network", "return_type"])
 def shreve(xp, river_network, return_type):
-    field = xp.zeros(river_network.n_nodes)
+    field = xp.zeros(river_network.n_nodes, dtype=int)
     # TODO: make xp-agnostic
     field[river_network.sources] = 1
